separated variants dowmload from variants views
